modify a InstructorBOImpl class

+++ src/main/java/lk/ijse/orm_coursework/bo/custom/impl/InstructorBOImpl.java
@@ -1,64 +1,77 @@
 package lk.ijse.orm_coursework.bo.custom.impl;
 
 import lk.ijse.orm_coursework.bo.custom.InstructorBO;
-import lk.ijse.orm_coursework.bo.exception.DuplicateException;
-import lk.ijse.orm_coursework.bo.exception.InUseException;
-import lk.ijse.orm_coursework.bo.util.EntityDTOConvertor;
+import lk.ijse.orm_coursework.bo.util.EntityDTOConverter;
 import lk.ijse.orm_coursework.dao.DAOFactory;
 import lk.ijse.orm_coursework.dao.DAOTypes;
 import lk.ijse.orm_coursework.dao.custom.InstructorDAO;
 import lk.ijse.orm_coursework.dto.InstructorDTO;
-import lk.ijse.orm_coursework.entity.Instructor;
+import lk.ijse.orm_coursework.entity.Instructors;
 
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+
 
 public class InstructorBOImpl implements InstructorBO {
 
-    private final InstructorDAO instructorDAO = DAOFactory.getInstance().getDAO(DAOTypes.INSTRUCTOR);
-    private final EntityDTOConvertor convertor = new EntityDTOConvertor();
+    private final InstructorDAO instructorDAO = DAOFactory.getInstance().getDAO(DAOTypes.INSTRUCTORS);
+    private final EntityDTOConverter convertor = new EntityDTOConverter();
 
 
     @Override
-    public void saveInstructor(InstructorDTO dto) throws DuplicateException, Exception {
-        Instructor instructor = convertor.getInstructor(dto);
-
-        instructorDAO.save(instructor);
-    }
-    @Override
-    public List<InstructorDTO> getAllInstructor() throws SQLException {
-        List<Instructor> instructors = instructorDAO.getAll();
-        List<InstructorDTO> dtos = new ArrayList<>();
-        for (Instructor instructor : instructors) {
-            dtos.add(convertor.getInstructorDTO(instructor));
+    public List<InstructorDTO> getAllInstructors() throws Exception {
+        List<Instructors> instructors = instructorDAO.getAll();
+        List<InstructorDTO> instructorDTOs = new ArrayList<>();
+        for (Instructors instructor : instructors) {
+            instructorDTOs.add(convertor.getInstructorsDTO(instructor));
         }
-        return dtos;
-    }
-
-
-
-    @Override
-    public void updateInstructor(InstructorDTO dto) throws SQLException {
-        Instructor instructor = convertor.getInstructor(dto);
-        instructorDAO.update(instructor);
+        return instructorDTOs;
     }
 
     @Override
-    public boolean deleteInstructor(String id) throws InUseException, Exception {
+    public String getLastInstructorId() throws Exception {
+        return instructorDAO.getLastId();
+    }
+
+    @Override
+    public boolean saveInstructors(InstructorDTO t) throws Exception {
+        Optional<Instructors> instructors = instructorDAO.findById(t.getInstructorId());
+        if (instructors.isPresent()) {
+            throw new Exception("Instructor already exists");
+        }
+        return instructorDAO.save(convertor.getInstructorsEntity(t));
+    }
+
+    @Override
+    public boolean updateInstructors(InstructorDTO t) throws Exception {
+        Optional<Instructors> instructors = instructorDAO.findById(t.getInstructorId());
+        if (instructors.isEmpty()) {
+            throw new Exception("Instructor Not Found");
+        }
+        return instructorDAO.update(convertor.getInstructorsEntity(t));
+    }
+
+    @Override
+    public boolean deleteInstructors(String id) throws Exception {
+        Optional<Instructors> instructors = instructorDAO.findById(id);
+        if (instructors.isEmpty()) {
+            throw new Exception("Instructor not Found");
+        }
         return instructorDAO.delete(id);
     }
 
     @Override
-    public String getNextId() throws SQLException {
-        String lastId = instructorDAO.getLastId();
-        char tableChar = 'I';
-        if (lastId != null){
-            String lastIdNumberString = lastId.substring(1);
-            int lastIdNumber = Integer.parseInt(lastIdNumberString);
-            int nextIdNumber = lastIdNumber + 1;
-            return String.format(tableChar + "%03d", nextIdNumber);
+    public List<String> getAllInstructorIds() throws Exception {
+        return instructorDAO.getAllIds();
+    }
+
+    @Override
+    public Optional<InstructorDTO> findByInstructorId(String id) throws Exception {
+        Optional<Instructors> instructors = instructorDAO.findById(id);
+        if (instructors.isPresent()) {
+            return Optional.of(convertor.getInstructorsDTO(instructors.get()));
         }
-        return tableChar + "001";
+        return Optional.empty();
     }
 }
